logs work in progress
