+++ android/src/main/java/io/parity/signer/screens/logs/LogsScreen.kt
@@ -4,20 +4,21 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import io.parity.signer.R
+import io.parity.signer.components.base.ScreenHeader
 import io.parity.signer.components.panels.BottomBar2
 import io.parity.signer.components.panels.BottomBar2State
 import io.parity.signer.models.Navigator
+import io.parity.signer.models.navigate
+import io.parity.signer.screens.logs.items.LogItem
+import io.parity.signer.screens.logs.items.LogItemDate
 import io.parity.signer.uniffi.Action
-import io.parity.signer.uniffi.Event
 import io.parity.signer.uniffi.History
-import io.parity.signer.uniffi.MLog
-import java.time.DayOfWeek
 
 
 @Composable
@@ -26,15 +27,17 @@ fun LogsScreen(
 	navigator: Navigator,
 ) {
 	Column(Modifier.background(MaterialTheme.colors.background)) {
-//	ScreenHeader(stringId = , backEnabled = , menuEnabled = , navigator = )
+		ScreenHeader(stringId = R.string.logs_title, onMenu = {
+			navigator.navigate(Action.RIGHT_BUTTON_ACTION)
+		})
 		Box() {
 			LazyColumn {
 				items(
 					items = model.logs,
 					key = { it.hashCode() }
 				) { item ->
-					when(item) {
-						is LogsListEntry.LogEntryModel -> {
+					when (item) {
+						is LogsListEntryModel.LogEntryModel -> {
 							Box(
 								Modifier.clickable {
 									navigator.navigate(
@@ -43,14 +46,13 @@ fun LogsScreen(
 									)
 								}
 							) {
-//							HistoryCard( //todo dmitry item create
-//								item,
-//								timestamp
-//							)
+								//todo dmitry
+//								LogItem(title = item., message =, time =)
 							}
 						}
-						is LogsListEntry.TimeSeparatorModel -> {
+						is LogsListEntryModel.TimeSeparatorModel -> {
 							//todo dmitry
+//							LogItemDate(date = item.)
 						}
 					}
 				}
@@ -60,43 +62,6 @@ fun LogsScreen(
 	}
 }
 
-@Composable
-fun LogItem(
-	model: History
-) {
-
-}
-
-data class LogsScreenModel(val logs: List<LogsListEntry>)
-
-fun MLog.toLogListEntries(): List<LogsListEntry> {
-	val logs: Sequence<LogsListEntry.LogEntryModel> =
-		log.asSequence().flatMap { it.toLogEntryModels() }
-	val result = mutableListOf<LogsListEntry>()
-
-	logs.forEach {
-		//todo dmitry add dates
-		result.add(it)
-	}
-	return result.toList()
-}
 
 
-sealed class LogsListEntry {
-	data class LogEntryModel(
-		val timestamp: String,
-		//id of this group of events, not unique per event
-		val logGroupId: UInt,
-		val event: Event,
-	) : LogsListEntry()
-
-	data class TimeSeparatorModel(
-		val month: String,
-		val dayOfWeek: Byte,
-		val year: Int,
-	) : LogsListEntry()
-}
-
-fun History.toLogEntryModels(): List<LogsListEntry.LogEntryModel> =
-	events.map { LogsListEntry.LogEntryModel(timestamp, order, it) }
