+++ android/src/main/java/io/parity/signer/ui/navigationselectors/SignerSelectorsLegacy.kt
@@ -9,6 +9,7 @@ import io.parity.signer.bottomsheets.*
 import io.parity.signer.components.Documents
 import io.parity.signer.models.*
 import io.parity.signer.screens.*
+import io.parity.signer.screens.logs.HistoryScreen
 import io.parity.signer.screens.logs.logdetails.LogDetails
 import io.parity.signer.uniffi.Action
 import io.parity.signer.uniffi.AlertData
