+++ android/src/main/res/values/strings.xml
@@ -71,4 +71,5 @@
 	<string name="log_title_warning">Warning!</string>
 	<string name="log_title_wrong_password">Wrong password entered</string>
 	<string name="log_message_wrong_passowrd">operation declined</string>
+	<string name="logs_title">All Logs</string>
 </resources>
